Adds support for Embrer 1.13
Fixes #16
